financial-document-analyzer-debug/requirements.txt
# Core Framework & Web Server
crewai==1.9.3
crewai-tools==1.9.3
fastapi==0.118.0
uvicorn==0.37.0
python-dotenv==1.1.1

# LLM & Language Processing
langchain-openai==1.1.10
langchain-community==0.4.1
langchain-core==1.2.15
openai==1.83.0
tiktoken==0.8.0

# PDF Processing
pypdf==6.7.3
pdfplumber==0.11.9

# Database (Bonus Feature)
sqlalchemy==2.0.45
# sqlite3 is built in, no need to install

# Async Processing & Message Queue (Bonus Feature)
celery==5.4.0
redis==5.2.0

# Utilities
pydantic==2.11.10
pydantic-settings==2.10.1
requests==2.32.5
aiohttp==3.13.2
numpy==2.3.5
